+++ src/app/componentes/proyectos/proyectos.component.ts
@@ -3,6 +3,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { Proyecto } from 'src/app/model/proyecto';
 import { ProyectoService } from 'src/app/service/proyecto.service';
 import { TokenService } from 'src/app/service/token.service';
+import { ConfirmComponent } from '../confirm/confirm.component';
 import { AddProyectoComponent } from './add-proyecto-component/add-proyecto.component';
 import { EditProyectoComponent } from './edit-proyecto-component/edit-proyecto.component';
 
@@ -42,12 +43,15 @@ export class ProyectosComponent implements OnInit {
   }
 
   delete(id:number){
-    if(id != undefined)
-      this.proyectoService.delete(id).subscribe(data => {
-        this.cargarProyectos();
-      }, err => {
-        alert("error al borrar id");
-      });
+    const dialogRef = this.dialog.open(ConfirmComponent, {maxWidth:'300px', width: '100%'});
+    dialogRef.afterClosed().subscribe(result => {
+      if(id != undefined && result)
+        this.proyectoService.delete(id).subscribe(data => {
+          this.cargarProyectos();
+        }, err => {
+          alert("error al borrar id");
+        });
+    })
   }
 
 }
